Adds country code for Android
`getCurrentLanguage` now returns languages in the format "xx-XX" where the first two 'x's are the language code and the second two are the country code.
This brings Android in line with iOS and ensures that language variants such as simplified or traditional Chinese aren't lost between the native app side and the react side.